Rectpacking: Moved OptimizationGoal enum to options package.
